# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Nuobuddy API",
    "description": "AI Agent Chatbot Backend API powered by Dify.ai. Provides user authentication, conversation management, and streaming chat integration with Dify.",
    "version": "1.0.0",
    "contact": {
      "name": "Nuobuddy Team"
    }
  },
...
  "paths": {
...
    "/chat": {
      "get": {
        "summary": "Chat API info",
        "deprecated": false,
        "description": "Returns information about the Chat API endpoints.",
        "tags": [
          "Chat"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Chat API information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/chat/conversations": {
      "get": {
        "summary": "Get user's conversation list",
        "deprecated": false,
        "description": "Returns a list of all conversations for the authenticated user, ordered by creation time (newest first).",
        "tags": [
          "Chat"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number for pagination",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of conversations per page",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Conversation list retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                "example": {
                  "status": 200,
                  "data": {
                    "conversations": [
                      {
                        "id": "550e8400-e29b-41d4-a716-446655440000",
                        "title": "Chat with NuoBuddy",
                        "difyConversationId": "dify-abc123",
                        "createdAt": "2026-03-22T10:00:00.000Z",
                        "updatedAt": "2026-03-22T10:30:00.000Z"
                      }
                    ],
                    "pagination": {
                      "page": 1,
                      "limit": 10,
                      "total": 1,
                      "totalPages": 1
                    }
                  },
                  "message": "Success"
                }
              }
            },
            "headers": {}
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "summary": "Create a new conversation",
        "deprecated": false,
        "description": "Creates a new conversation for the authenticated user.",
        "tags": [
          "Chat"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConversationRequest"
              },
              "example": {
                "title": "测试对话1"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Conversation created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                "example": {
                  "status": 200,
                  "data": {
                    "id": "550e8400-e29b-41d4-a716-446655440000",
                    "title": "New Chat",
                    "difyConversationId": null,
                    "createdAt": "2026-03-22T10:00:00.000Z"
                  },
                  "message": "Success"
                }
              }
            },
            "headers": {}
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/chat/deleteConversations": {
      "post": {
        "summary": "Delete a conversation",
        "deprecated": false,
        "description": "Deletes a conversation and its associated messages. Only the conversation owner can delete it.",
        "tags": [
          "Chat"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Conversation deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            },
            "headers": {}
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          },
          "403": {
            "description": "Forbidden - not the conversation owner",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          },
          "404": {
            "description": "Conversation not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/chat/conversations/{id}/message": {
      "post": {
        "summary": "Send a chat message (SSE streaming)",
        "deprecated": false,
        "description": "Sends a message to the Dify AI agent and receives a streaming response via Server-Sent Events. The response is piped directly from Dify to the client in real-time.",
        "tags": [
          "Chat"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The conversation ID (UUID). Use empty string for new conversation.",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageRequest"
              },
              "examples": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "SSE stream established",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/SSEStreamResponse"
                }
              }
            },
            "headers": {}
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/chat/conversations/{id}/share": {
      "post": {
        "summary": "Create share link for conversation",
        "deprecated": false,
        "description": "Creates a shareable link for the conversation by setting the share field to true. Requires authentication and conversation ownership.",
        "tags": [
          "Chat"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The conversation ID (UUID)",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Share link created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                "example": {
                  "status": 200,
                  "data": {
                    "shareUrl": "https://example.com/chat/550e8400-e29b-41d4-a716-446655440000?share=1"
                  },
                  "message": "Share link created successfully"
                }
              }
            },
            "headers": {}
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          },
          "403": {
            "description": "Forbidden - not the conversation owner",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          },
          "404": {
            "description": "Conversation not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "summary": "Cancel share link for conversation",
        "deprecated": false,
        "description": "Cancels the shareable link for the conversation by setting the share field to false. Requires authentication and conversation ownership.",
        "tags": [
          "Chat"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The conversation ID (UUID)",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Share link cancelled successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                "example": {
                  "status": 200,
                  "data": null,
                  "message": "Share link cancelled successfully"
                }
              }
            },
            "headers": {}
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          },
          "403": {
            "description": "Forbidden - not the conversation owner",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          },
          "404": {
            "description": "Conversation not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "summary": "Get conversation with share status",
        "deprecated": false,
        "description": "Retrieves a conversation with its messages. When share=1 query parameter is provided, checks if the conversation is shared and validates access permissions. Does not require authentication for shared links.",
        "tags": [
          "Chat"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The conversation ID (UUID)",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "share",
            "in": "query",
            "description": "Set to '1' to access as shared link (no authentication required)",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["1"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Conversation retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                "example": {
                  "status": 200,
                  "data": {
                    "id": "550e8400-e29b-41d4-a716-446655440000",
                    "title": "Chat with NuoBuddy",
                    "difyConversationId": "dify-abc123",
                    "createdAt": "2026-03-22T10:00:00.000Z",
                    "updatedAt": "2026-03-22T10:30:00.000Z",
                    "messages": [
                      {
                        "id": "msg-1",
                        "role": "user",
                        "content": "Hello!",
                        "timestamp": "2026-03-22T10:00:00.000Z"
                      },
                      {
                        "id": "msg-2",
                        "role": "assistant",
                        "content": "Hi there!",
                        "timestamp": "2026-03-22T10:00:05.000Z"
                      }
                    ],
                    "share": true,
                    "accessible": true
                  },
                  "message": "Success"
                }
              }
            },
            "headers": {}
          },
          "401": {
            "description": "Unauthorized (required for non-shared access)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          },
          "403": {
            "description": "Forbidden - conversation not shared or access denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "status": 403,
                  "data": {
                    "share": false,
                    "accessible": false
                  },
                  "message": "This conversation is not shared or you don't have access"
                }
              }
            },
            "headers": {}
          },
          "404": {
            "description": "Conversation not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "description": "HTTP status code"
          },
          "data": {
            "type": "object",
            "description": "Response data",
            "properties": {},
            "nullable": true
          },
          "message": {
            "type": "string",
            "description": "Response message"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "description": "HTTP status code"
          },
          "data": {
            "type": "null",
            "nullable": true
          },
          "message": {
            "type": "string",
            "description": "Error message"
          }
        }
      },
...
      "CreateConversationRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Optional conversation title"
          }
        }
      },
      "SendMessageRequest": {
        "type": "object",
        "required": [
          "query"
        ],
        "properties": {
          "query": {
            "type": "string",
            "description": "The user's message to the AI agent"
          }
        }
      },
...
      "SSEStreamResponse": {
        "type": "string",
        "description": "Server-Sent Events stream with the following event types:\n- `delta`: Partial AI response content\n- `done`: Stream completed with metadata\n- `error`: Error occurred\n- `ping`: Heartbeat (can be ignored)",
        "example": "event: delta\ndata: {\"content\": \"Hello!\"}\n\nevent: done\ndata: {\"conversationId\": \"550e8400-e29b-41d4-a716-446655440000\"}\n\n"
      },
...
    },
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /user/login"
      }
    }
  },
  "servers": [],
  "security": [
    {
      "bearerAuth": []
    }
  ]
}